Back

See merge request s09-webmobile1-sub2/S09P12B304!238

+++ Back/src/main/java/com/ssafy/howdoilook/domain/user/service/UserService.java
@@ -241,7 +241,7 @@ public class UserService {
 
         String imageUrl = userUpdateProfileImgDto.getImageUrl();
 
-        if(extractFileNameFromUrl(imageUrl).equals("DefaultProfile")) {
+        if(extractFileNameFromUrl(imageUrl).equals("DefaultProfile.png")) {
             user.updateProfileImg(imageService.saveImage(multipartFile));
             return UserUpdateProfileImgResponseDto.builder()
                     .profileImg(user.getProfileImg())
@@ -269,6 +269,10 @@ public class UserService {
 
         if (!clientEmail.equals(user.getEmail())){
             throw new AccessException("접근 권한이 없습니다.");
+        }
+
+        if(extractFileNameFromUrl(user.getProfileImg()).equals("DefaultProfile.png")) {
+            return;
         }
 
         imageService.deleteImage(user.getProfileImg());
